+++ andes/models/experimental/__init__.py
@@ -5,3 +5,4 @@ Experimental models.
 from andes.models.experimental.fixed import FixedGen  # NOQA
 from andes.models.experimental.picontroller import PI2  # NOQA
 from andes.models.experimental.testmodel import TestDB1, TestPI, TestLagAWFreeze  # NOQA
+from andes.models.experimental.genrouos import GENROUOS  # NOQA
