--- andes/models/sync/__init__.py
@@ -1,5 +1,0 @@
-"""
-Package for synchronous generators with operator splitting
-"""
-
-from andes.models.sync.genrou import GENROUOS  # NOQA
